♻️ 요소 컴포넌트 분리 ( #97)
