+++ src/pages/recommendation-proposal/index.jsx
@@ -1,3 +1,7 @@
 import ProposalRecPage from './ProposalRecPage';
+import BigCard from './Card/BigCard';
+import SmallCard from './Card/SmallCard';
 
 export default ProposalRecPage;
+
+export { BigCard, SmallCard };
